+++ src/main/java/com/verdantartifice/primalmagick/common/theorycrafting/TheorycraftSpeed.java
@@ -1,0 +1,7 @@
+package com.verdantartifice.primalmagick.common.theorycrafting;
+
+public enum TheorycraftSpeed {
+    SLOW,
+    NORMAL,
+    FAST;
+}